+++ src/features/swap/estimateGasFee.ts
@@ -1,10 +1,11 @@
-import BigNumber from "bignumber.js";
 import { SUI_DECIMALS } from "@mysten/sui/utils";
-import { buildTx } from "./buildTx";
-import { formatBalance } from "../../utils/number";
-import { EstimateGasFeeParams } from "../../types/tx";
-import { getSuiPrice } from "../prices";
 import { Config } from "../../config";
+import { isBluefinXRouting } from "../../types/aggregator";
+import { EstimateGasFeeParams } from "../../types/tx";
+import { formatBalance } from "../../utils/number";
+import { getSuiPrice } from "../prices";
+import { buildTx } from "./buildTx";
+import { BluefinXTx } from "../../libs/protocols/bluefinx/types";
 
 export async function estimateGasFee({
   quoteResponse,
@@ -15,6 +16,8 @@ export async function estimateGasFee({
   commission,
 }: EstimateGasFeeParams): Promise<number> {
   if (!accountAddress) return 0;
+  // BluefinX is sponsored, no need to estimate gas fee
+  if (!accountAddress || isBluefinXRouting(quoteResponse)) return 0;
 
   const result = await buildTx({
     extendTx,
@@ -30,7 +33,7 @@ export async function estimateGasFee({
 
   const { tx } = result || {};
 
-  if (!tx) return 0;
+  if (!tx || tx instanceof BluefinXTx) return 0;
 
   const suiPrice = _suiPrice || (await getSuiPrice());
   const suiDecimals = SUI_DECIMALS;
@@ -44,12 +47,11 @@ export async function estimateGasFee({
 
   if (status.status !== "success") return 0;
 
-  const fee = new BigNumber(gasUsed.computationCost)
-    .plus(gasUsed.storageCost)
-    .minus(gasUsed.storageRebate);
-  const feeUsd = new BigNumber(suiPrice).multipliedBy(
-    formatBalance(fee, suiDecimals),
-  );
+  const fee =
+    BigInt(gasUsed.computationCost) +
+    BigInt(gasUsed.storageCost) -
+    BigInt(gasUsed.storageRebate);
+  const feeUsd = Number(suiPrice) * Number(formatBalance(fee, suiDecimals));
 
-  return feeUsd.toNumber();
+  return feeUsd;
 }