+++ src/index.ts
@@ -2,26 +2,29 @@ export * from "./types/aggregator";
 
 import { Config } from "./config";
 
-import { getTokenPrice, getTokenPrices, getSuiPrice } from "./features/prices";
+import { getSuiPrice, getTokenPrice, getTokenPrices } from "./features/prices";
+import { executeBluefinTx } from "./libs/protocols/bluefinx/client";
 
 import {
   buildTx,
-  getQuote,
   estimateGasFee,
+  executeTx,
+  getQuote,
   getSwapHistory,
+  DEFAULT_SOURCES,
 } from "./features/swap";
 
 import {
-  placeLimitOrder,
-  getOpenLimitOrders,
+  cancelDcaOrder,
   cancelLimitOrder,
   claimExpiredLimitOrder,
-  getClosedLimitOrders,
-  placeDcaOrder,
-  getOpenDcaOrders,
-  cancelDcaOrder,
   getClosedDcaOrders,
+  getClosedLimitOrders,
   getDcaOrderExecutions,
+  getOpenDcaOrders,
+  getOpenLimitOrders,
+  placeDcaOrder,
+  placeLimitOrder,
 } from "./features/limitDca";
 
 // avoid breaking changes
@@ -46,6 +49,9 @@ export {
   estimateGasFee,
   buildTx,
   getSwapHistory,
+  executeTx,
+  executeBluefinTx,
+  DEFAULT_SOURCES,
 
   // limit order
   placeLimitOrder,
@@ -80,6 +86,9 @@ export default {
   estimateGasFee,
   buildTx,
   getSwapHistory,
+  executeTx,
+  executeBluefinTx,
+  DEFAULT_SOURCES,
 
   // limit order
   placeLimitOrder,