+++ src/libs/protocols/base.ts
@@ -9,24 +9,29 @@ export interface BaseContractParams {
   inputCoinObject: TransactionResultItem;
   currentAccount: string;
   config: Config;
+  /** map price feed id to onchain priceInfoObject id */
+  pythMap: Record<string, string>;
 }
 
-export abstract class BaseContract {
+export abstract class BaseContract<T = any> {
   protected swapInfo: TxSorSwap;
   protected inputCoinObject: TransactionResultItem;
   protected currentAccount: string;
   protected config: Config;
+  protected pythMap: Record<string, string>;
 
   constructor({
     swapInfo,
     inputCoinObject,
     currentAccount,
     config,
+    pythMap,
   }: BaseContractParams) {
     this.swapInfo = swapInfo;
     this.inputCoinObject = inputCoinObject;
     this.currentAccount = currentAccount;
     this.config = config;
+    this.pythMap = pythMap;
   }
 
   abstract swap(tx: Transaction): Promise<TransactionResultItem>;
@@ -43,5 +48,13 @@ export abstract class BaseContract {
     return parseStructTag(
       this.swapInfo.extra?.poolStructTag || "",
     ).typeParams.map(normalizeStructTag);
+  }
+
+  protected get extra() {
+    const extra = this.swapInfo.extra as T;
+    if (!extra) {
+      throw new Error(`Invalid extra info for getExtra`);
+    }
+    return extra;
   }
 }